Organised project directory structure
Improve seperation of concern

See #2
